secure_mvp/backend/requirements.txt
Django==5.0.14
djangorestframework==3.15.2
psycopg[binary]==3.2.6
django-cors-headers==4.6.0
Pillow==11.1.0
